Renamed weekplans_bloc to weekplan_selector_bloc
